port SHA-1: 2b48aec2c43c105fe128e3998792f56204a9d32a
* Possible fix for https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=37811

+++ Source/lib/aztec/encoder/BinaryShiftToken.cs
@@ -25,8 +25,8 @@ namespace ZXing.Aztec.Internal
     /// </summary>
     public sealed class BinaryShiftToken : Token
     {
-        private readonly short binaryShiftStart;
-        private readonly short binaryShiftByteCount;
+        private readonly int binaryShiftStart;
+        private readonly int binaryShiftByteCount;
 
         /// <summary>
         /// initializing constructor
@@ -39,8 +39,8 @@ namespace ZXing.Aztec.Internal
                                 int binaryShiftByteCount)
            : base(previous)
         {
-            this.binaryShiftStart = (short)binaryShiftStart;
-            this.binaryShiftByteCount = (short)binaryShiftByteCount;
+            this.binaryShiftStart = binaryShiftStart;
+            this.binaryShiftByteCount = binaryShiftByteCount;
         }
 
         /// <summary>
